backend/requirements.txt
# Web Framework
Flask==2.3.3
gunicorn==23.0.0
Werkzeug==3.1.3

# API Communication
requests==2.32.3
Flask-Cors==6.0.1

# Environment Variables
python-dotenv==1.1.0

# Other Dependencies
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2